Error handling codes added

+++ controller/contacts.controller.js
@@ -1,16 +1,30 @@
 import Contacts from "../models/contacts.models.js";
+import mongoose from "mongoose";
 
 export const getFindContacts = async (req, res) => {
   // home page
-  const contacts = await Contacts.find();
-  // res.json(contacts); //shows data in json format
-  res.render("home", { contacts });
+  try {
+    const contacts = await Contacts.find();
+    if (!contacts) return res.render("404", { message: "Not data found" });
+    res.render("home", { contacts });
+  } catch (error) {
+    res.render("500", { message: error });
+  }
 };
 
 export const getFindContactById = async (req, res) => {
   //static page to show contact details
-  const contact = await Contacts.findOne({ _id: req.params.id });
-  res.render("show-contact", { contact });
+  const validId = await mongoose.Types.ObjectId.isValid(req.params.id);
+  if (!validId) return res.render("404", { message: "Invalid ID" });
+
+  try {
+    const contact = await Contacts.findOne({ _id: req.params.id });
+
+    if (!contact) return res.render("404", { message: "Contact not Found" });
+    res.render("show-contact", { contact });
+  } catch (error) {
+    res.render("500", { message: error });
+  }
 };
 
 export const getAddContact = (req, res) => {
@@ -20,35 +34,61 @@ export const getAddContact = (req, res) => {
 export const postAddContact = async (req, res) => {
   //post the contact input into db
   // const contact_2= Contacts.insertOne({
-  //   first_name: first_name,
-  //   last_name: last_name,
   //   "db_field_name": "form_field_name"
   // }) //but this can be done in a shorter and simpler way.
 
-  const contact = await Contacts.create(req.body);
-  // in this case "db_field_name" and "form_field_name" should be the same
-  //if i'm not using the const "contact anywhere else", directly "await Contacts.create(req.body)" can also be written
-  res.redirect("/");
+  try {
+    const contact = await Contacts.create(req.body);
+    if (!contact)
+      return res.render("500", { message: "Could not insert data" });
+    // in this case "db_field_name" and "form_field_name" should be the same
+    res.redirect("/");
+  } catch (error) {
+    res.render("500", { message: error });
+  }
 };
 
 export const getUpdateContactById = async (req, res) => {
   //form page to take contact update from user
-  const contact = await Contacts.findOne({ _id: req.params.id });
-  res.render("update-contact", { contact });
+  const validId = await mongoose.Types.ObjectId.isValid(req.params.id);
+  if (!validId) return res.render("404", { message: "Invalid ID" });
+
+  try {
+    const contact = await Contacts.findOne({ _id: req.params.id });
+    if (!contact) return res.render("404", { message: "Contact Not Found" });
+    res.render("update-contact", { contact });
+  } catch (error) {
+    res.render("500", { message: error });
+  }
 };
 
 export const postUpdateContactById = async (req, res) => {
   //post the contact update into db
   const id = req.params.id;
   const updateData = req.body;
-  const updateContact = await Contacts.findByIdAndUpdate(id, updateData);
-  // const updateContact= await Contacts.findByIdAndUpdate(req.params.id, req.body) //in short
+  const validId = await mongoose.Types.ObjectId.isValid(id);
+  if (!validId) return res.render("404", { message: "Invalid ID" });
 
-  res.redirect("/");
+  try {
+    const updateContact = await Contacts.findByIdAndUpdate(id, updateData);
+    if (!updateContact)
+      return res.render("404", { message: "Contact Not Found" });
+    res.redirect("/");
+  } catch (error) {
+    res.render("500", { message: error });
+  }
 };
 
 export const deleteContactById = async (req, res) => {
   //contact will be deleted from db
-  const deleteData = await Contacts.findByIdAndDelete(req.params.id);
-  res.redirect("/");
+  const validId = await mongoose.Types.ObjectId.isValid(req.params.id);
+  if (!validId) return res.render("404", { message: "Invalid ID" });
+
+  try {
+    const deleteData = await Contacts.findByIdAndDelete(req.params.id);
+    if (!deleteData) return res.render("404", { message: "Contact Not Found" });
+    res.redirect("/");
+  } catch (error) {
+    res.render("500", { message: error });
+  }
 };
